requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
python-dotenv==1.0.0
serpapi==0.1.0
gunicorn==21.2.0
python-socketio==5.9.0
